Create README.md
Problem Explanation
Instead of using string concatenation, ES6 offers template literals to create strings. In this challenge, you have to use template literals to create an array of text warnings.

It’s required to use template literals to return a list as every element in the array as the element will be wrapped in a <li></li> tag.

Hints
Hint 1
Iterate through each element of arr and add a new element consisting composed of a template literal using arr's elements to create the array failureItems.